backend: move errorresult to util package
