+++ src/backend/controllers/base/logged.go
@@ -9,6 +9,7 @@ import (
 	"github.com/Qihoo360/wayne/src/backend/bus"
 	"github.com/Qihoo360/wayne/src/backend/bus/message"
 	"github.com/Qihoo360/wayne/src/backend/models"
+	"github.com/Qihoo360/wayne/src/backend/util/errors"
 	"github.com/Qihoo360/wayne/src/backend/util/hack"
 	"github.com/Qihoo360/wayne/src/backend/util/logs"
 	"github.com/dgrijalva/jwt-go"
@@ -45,7 +46,7 @@ func (c *LoggedInController) Prepare() {
 		return rsakey.RsaPublicKey, nil
 	})
 
-	errResult := ErrorResult{}
+	errResult := errors.ErrorResult{}
 	switch err.(type) {
 	case nil: // no error
 		if !token.Valid { // but may still be invalid
